+++ ChangeLog.md
@@ -7,6 +7,7 @@ This document is created to record all changes.
 - Add the header file `includes/ObserverPattern.h` to define the class `myAbstractObserver` for observer pattern.
 - Define the class `mySubjectInterface` for observer pattern.
 - Implement all member methods of the class `mySubjectInterface`.
+- Change the method `mySubjectInterface::notifyData()` from an abstract method to a instance method, and then implement it.
 
 ### Changed
 - Rename the class from `myAbstractObserver` to `myObserverInterface`.